agente_busquedas_externas/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agente-busquedas-externas"
version = "0.1.0"
description = "Multi-agent recruiting/candidate-sourcing system"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "a2a-sdk>=0.3.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
